refact: move browser-related things into browser.rs
